Split the VarianceColumn file into three files, and remove the dependence on categorySelection.py.
