+++ States/State.py
@@ -1,6 +1,5 @@
 import os
 from util import fileio
-from util import categorySelection as cate
 from util import readConfig as rc
 from util import timeHelpers as timeh
 from DataClasses import BptList
